apps/simple-clustering/requirements.txt
streamlit
torch
torchvision
pillow
numpy
matplotlib
scikit-learn
scipy
requests
